app/requirements.txt
Requests
Streamlit
Pillow